Imported configparser module in TicketViewer.py
Since we are not allowed to have a login page, I will use a separate config file for the end user to put their information in. This will allow the end user to easily enter their information without the need for a separate login state or requiring the user to hardcode their information into the program.